Basic setup of redux stuff, but man do i not know what i'm doing
