+++ src/actions/actionCreators.js
@@ -1,0 +1,5 @@
+export function toggleNewClientDrawer() {
+    return {
+        type: 'TOGGLE_NEW_CLIENT_DRAWER'
+    };
+}
