+++ src/components/Main/Main.scss
