+++ src/index.js
@@ -3,6 +3,53 @@ import { render } from 'react-dom';
 
 import './sass/style.css';
 
+// Import components
 import App from './components/App/App';
 
-render(<App />, document.getElementById('root'));
+// Redux and react things
+import { Provider } from 'react-redux';
+import { BrowserRouter, Switch, Route, Router, Link } from 'react-router-dom';
+import store from './store';
+
+// <Provider store={store}>
+//     <BrowserRouter>
+//         <App>
+//             <Switch>
+//                 <Route
+//                     path="/client/:clientSlug"
+//                     render={() => <h2>This is a client page</h2>}
+//                 />
+//                 <Route path="/" render={() => <p>No client page</p>} />
+//             </Switch>
+//         </App>
+//     </BrowserRouter>
+// </Provider>
+
+const router = (
+    <Provider store={store}>
+        <App />
+        {/* <App> */}
+        {/* <BrowserRouter>
+                <Switch>
+                    <Route
+                        path="/client/:clientSlug"
+                        component={() => (
+                            <h1>
+                                Client page <Link to="/">Home</Link>
+                            </h1>
+                        )}
+                    />
+                    <Route
+                        component={() => (
+                            <p>
+                                Homepage <Link to="/client/adam">Adam</Link>
+                            </p>
+                        )}
+                    />
+                </Switch>
+            </BrowserRouter> */}
+        {/* </App> */}
+    </Provider>
+);
+
+render(router, document.getElementById('root'));
