+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import { routerReducer } from 'react-router-redux';
+
+import view from './view';
+
+const rootReducer = combineReducers({ view, routing: routerReducer });
+
+export default rootReducer;
