+++ src/store.js
@@ -1,0 +1,27 @@
+import { createStore, compose } from 'redux';
+// import { syncHistoryWithStore } from 'react-router-redux';
+// import { browserHistory } from 'react-router';
+
+// import the route reducer
+import rootReducer from './reducers/index';
+
+// create an object for default data
+const defaultState = {
+    view: {
+        new_client_drawer: false
+    }
+};
+
+const enhancers = compose(window.devToolsExtension ? window.devToolsExtension() : f => f);
+
+const store = createStore(rootReducer, defaultState, enhancers);
+// export const history = syncHistoryWithStore(browserHistory, store);
+
+if (module.hot) {
+    module.hot.accept('./reducers/', () => {
+        const nextRootReducer = require('./reducers/index').default;
+        store.replaceReducer(nextRootReducer);
+    });
+}
+
+export default store;
